Extract log-request-console middleware
